flask_blog_api/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
psycopg2-binary
python-dotenv
